third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/image-as-flexitem-size-005v-expected.txt

This is a testharness.js-based test.
FAIL .flexbox > img 1 assert_equals: 
<img src="support/solidblue.png" data-expected-width="40" data-expected-height="40">
height expected 40 but got 16
FAIL .flexbox > img 2 assert_equals: 
<img src="support/solidblue.png" style="width: 30px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 30
PASS .flexbox > img 3
FAIL .flexbox > img 4 assert_equals: 
<img src="support/solidblue.png" style="flex: 1 1 30px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 16
FAIL .flexbox > img 5 assert_equals: 
<img src="support/solidblue.png" style="min-width: 34px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 34
FAIL .flexbox > img 6 assert_equals: 
<img src="support/solidblue.png" style="min-height: 34px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 34
FAIL .flexbox > img 7 assert_equals: 
<img src="support/solidblue.png" style="min-width: 30px;
                                              min-height: 34px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 34
FAIL .flexbox > img 8 assert_equals: 
<img src="support/solidblue.png" style="min-width: 34px;
                                              min-height: 30px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 34
PASS .flexbox > img 9
PASS .flexbox > img 10
PASS .flexbox > img 11
PASS .flexbox > img 12
PASS .flexbox > img 13
PASS .flexbox > img 14
PASS .flexbox > img 15
PASS .flexbox > img 16
PASS .flexbox > img 17
FAIL .flexbox > img 18 assert_equals: 
<img src="support/solidblue.png" style="width: 10px;
                                              min-height: 30px" data-expected-width="40" data-expected-height="40">
height expected 40 but got 30
Harness: the test ran to completion.
